Move UIExtensions content to ControlsHelper
ControlsHelper now owns LabelTextProperty
